ai/refactor: 依存関係解析をgraphパッケージからanalyzerパッケージに移動し責務分離を実現
